+++ src/main/java/com/project/baedalsodae/event/handler/OrderEventHandler.java
@@ -5,7 +5,7 @@ import com.project.baedalsodae.event.entity.AggregateType;
 import com.project.baedalsodae.event.entity.Event;
 import com.project.baedalsodae.event.entity.EventType;
 import com.project.baedalsodae.event.repository.EventRepository;
-import com.project.baedalsodae.order.dto.event.OrderCreatedEvent;
+import com.project.baedalsodae.event.dto.OrderCreatedEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
