+++ src/main/java/com/project/baedalsodae/payment/listener/PaymentListener.java
@@ -1,6 +1,6 @@
 package com.project.baedalsodae.payment.listener;
 
-import com.project.baedalsodae.order.dto.event.OrderCreatedEvent;
+import com.project.baedalsodae.event.dto.OrderCreatedEvent;
 import com.project.baedalsodae.payment.service.PaymentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
